scinsight: Try to match on log path as dir
In scinsight, when matching trying to find log files emitted by scmon
according to the --log argument, we do a hard match on
/path/to/log-scmon*. It seems like it would be a somewhat typical /
intuitive use case to have scmon emit logs to some directory, and for a
user to want to just pass a file to the directory and have scinsight
find them automatically.

Let's update get_strace_names() to also try to match by log-directory
prefix. Note that this will break users who always output to the same
output directory, but with different log file prefixes.

Signed-off-by: David Vernet <void@manifault.com>